Implemented dynamic menu, geolocations and other changes
