+++ FrontEnd/index.html
@@ -44,75 +44,7 @@
   </head>
   <body>
     <!-- <div id='x-body' class='container-full'> -->
-    <div id="x-header">
-      <nav class="navbar navbar-expand navbar-dark fixed-top bg-dark p-1">
-        <div class="container-fluid">
-          <img
-            src="https://www.asthrasoft.com/site/img/logo.png"
-            style="height: 25px"
-          />
-          <p class="navbar-brand my-0 ms-2">Think Nutrition</p>
-
-          <button
-            class="navbar-toggler"
-            type="button"
-            data-bs-toggle="collapse"
-            data-bs-target="#navbarCollapse"
-            aria-controls="navbarCollapse"
-            aria-expanded="false"
-            aria-label="Toggle navigation"
-          >
-            <span class="navbar-toggler-icon"></span>
-          </button>
-          <div class="collapse navbar-collapse" id="navbarCollapse">
-            <ul class="navbar-nav ms-auto">
-              <li class="nav-item">
-                <a
-                  class="nav-link active"
-                  aria-current="page"
-                  data-nav="admin.dashboard"
-                  href="/"
-                  >Dashboard</a
-                >
-              </li>
-              <li class="nav-item dropdown">
-                <a
-                  class="nav-link dropdown-toggle"
-                  href="#"
-                  id="dropdown03"
-                  data-bs-toggle="dropdown"
-                  aria-expanded="false"
-                  ><i class="bi bi-gear"></i
-                ></a>
-                <ul
-                  class="dropdown-menu dropdown-menu-end"
-                  aria-labelledby="dropdown03"
-                >
-                  <li>
-                    <a class="dropdown-item" href="/ngos" data-nav="admin.ngos"
-                      >NGOs</a
-                    >
-                  </li>
-                  <li>
-                    <a class="dropdown-item" href="/Aanganwadis">Aanganwadis</a>
-                  </li>
-                  <li><hr class="dropdown-divider" /></li>
-                  <li>
-                    <a class="dropdown-item" data-nav="settings">Settings</a>
-                  </li>
-                  <li>
-                    <a class="dropdown-item" data-nav="profile">Profile</a>
-                  </li>
-                  <li>
-                    <a class="dropdown-item" data-nav="web.login">Logout</a>
-                  </li>
-                </ul>
-              </li>
-            </ul>
-          </div>
-        </div>
-      </nav>
-    </div>
+ 
     <div id="app"></div>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
@@ -132,14 +64,7 @@
     ></script>
 
     <script type="module" src="/src/main.js"></script>
-    <footer
-      class="navbar navbar-expand navbar-dark fixed-bottom bg-dark p-1"
-      style="height: 30px"
-    >
-      <div class="container-fluid1">
-        <p class="text-light mt-3">&copy; Powered by Thinkar Labs</p>
-      </div>
-    </footer>
+  
     <!-- </div> -->
   </body>
 </html>
