+++ FrontEnd/src/App.vue
@@ -1,7 +1,7 @@
-<script setup>
-import { RouterLink, RouterView } from "vue-router";
-</script>
+<template><HeaderView></HeaderView> <RouterView /><FooterView></FooterView></template>
 
-<template>
-  <RouterView />
-</template>
+<script setup>
+import { RouterView } from "vue-router";
+import HeaderView from "../src/components/HeaderView.vue";
+import FooterView from "../src/components/FooterView.vue";
+</script>
