+++ FrontEnd/src/components/FooterView.vue
@@ -1,0 +1,10 @@
+<template>
+  <footer
+    class="navbar navbar-expand navbar-dark fixed-bottom bg-dark p-1"
+    style="height: 30px"
+  >
+    <div class="container-fluid1">
+      <p class="text-light mt-3">&copy; Powered by Thinkar Labs</p>
+    </div>
+  </footer>
+</template>
