+++ FrontEnd/src/views/Admin/EditNgoView.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="full-div container" style="width: 1280px">
     <div id="x-contest" class="container-fluid p-3">
-      <form v-on:submit.prevent="updateNgo">
+      <form @submit.prevent="updateNgo">
         <div class="row">
           <h3 class="float-start">Manage NGO</h3>
 
@@ -14,9 +14,6 @@
               placeholder="Ngo Name"
               v-model="updatedNgo.ngoName"
             />
-            <div className="text-danger mrgnbtn" v-if="helperSupport.ngoName">
-              {{ helperSupport.ngoName }}
-            </div>
           </div>
 
           <div class="col-6 my-2">
@@ -28,12 +25,6 @@
               placeholder="Contact Person Name"
               v-model="updatedNgo.contactPersonName"
             />
-            <div
-              className="text-danger mrgnbtn"
-              v-if="helperSupport.contactPersonName"
-            >
-              {{ helperSupport.contactPersonName }}
-            </div>
           </div>
 
           <div class="col-6 my-2">
@@ -45,12 +36,6 @@
               placeholder="Contact Person Phone"
               v-model="updatedNgo.contactPersonPhone"
             />
-            <div
-              className="text-danger mrgnbtn"
-              v-if="helperSupport.contactPersonName"
-            >
-              {{ helperSupport.contactPersonName }}
-            </div>
           </div>
 
           <div class="col-6 my-2">
@@ -62,18 +47,10 @@
               placeholder="Contact Person Email"
               v-model="updatedNgo.contactPersonEmail"
             />
-            <div
-              className="text-danger mrgnbtn"
-              v-if="helperSupport.contactPersonEmail"
-            >
-              {{ helperSupport.contactPersonEmail }}
-            </div>
           </div>
 
           <div class="col-6 my-2">
-            <label for="exampleFormControlInput1"
-              >Contact Person Password</label
-            >
+            <label for="exampleFormControlInput1">Contact Person Password</label>
             <input
               type="password"
               class="form-control"
@@ -81,12 +58,6 @@
               placeholder="Password"
               v-model="updatedNgo.contactPersonPassword"
             />
-            <div
-              className="text-danger mrgnbtn"
-              v-if="helperSupport.contactPersonPassword"
-            >
-              {{ helperSupport.contactPersonPassword }}
-            </div>
           </div>
           <div class="col-6 my-2">
             <label for="exampleFormControlInput1">Location</label>
@@ -97,9 +68,6 @@
               placeholder="Location"
               v-model="updatedNgo.location"
             />
-            <div className="text-danger mrgnbtn" v-if="helperSupport.location">
-              {{ helperSupport.location }}
-            </div>
           </div>
           <div class="col-6 my-2">
             <label for="exampleFormControlInput1">Pincode</label>
@@ -110,9 +78,6 @@
               placeholder="pincode"
               v-model="updatedNgo.pincode"
             />
-            <div className="text-danger mrgnbtn" v-if="helperSupport.pincode">
-              {{ helperSupport.pincode }}
-            </div>
           </div>
           <div class="row">
             <div class="col-12 p-2">
@@ -169,17 +134,16 @@ import { ref, onMounted, computed, reactive } from "vue";
 import { useNgoStore } from "../stores/ngo";
 import { useRoute } from "vue-router";
 import router from "../router";
-import helper from "../helper/validation.helper.js";
 
 let updatedNgo = reactive({
   id: "",
   ngoName: "",
   contactPersonName: "",
   contactPersonEmail: "",
-  contactPersonPhone: "",
+  contactPersonPhone: 0,
   contactPersonPassword: "",
   location: "",
-  pincode: "",
+  pincode: 0,
 });
 
 const route = useRoute();
@@ -201,37 +165,6 @@ updatedNgo = computed(() => {
     };
   }
 });
-
-const helperSupport = reactive({
-  ngoName: "",
-  contactPersonName: "",
-  contactPersonEmail: "",
-  contactPersonPhone: "",
-  contactPersonPassword: "",
-  location: "",
-  pincode: "",
-});
-
-const isValidSubmission = (updatedNgo) => {
-  helperSupport.pincode = helper.validatePincode(updatedNgo._value.pincode);
-  helperSupport.contactPersonName = helper.validateName(
-    updatedNgo._value.contactPersonName
-  );
-  helperSupport.contactPersonEmail = helper.validateEmail(
-    updatedNgo._value.contactPersonEmail
-  );
-  helperSupport.ngoName = helper.validateName(updatedNgo._value.ngoName);
-  helperSupport.contactPersonPhone = helper.validatePhoneNumber(
-    updatedNgo._value.contactPersonPhone
-  );
-  helperSupport.location = helper.validateName(updatedNgo._value.location);
-  helperSupport.contactPersonPassword =
-    updatedNgo._value.contactPersonPassword !== ""
-      ? ""
-      : "Password is mandatory";
-  return helper.isErrorMessagesAvailable(helperSupport) ? false : true;
-};
-
 onMounted(async () => {
   console.log("Editing.... ");
   console.log(route.params.id);
@@ -240,9 +173,7 @@ onMounted(async () => {
 });
 
 const updateNgo = async () => {
-  if (isValidSubmission(updatedNgo) == true) {
-    await store.updateNgo(updatedNgo);
-    return router.push("/ngos");
-  }
+  await store.updateNgo(updatedNgo);
+  return router.push("/ngos");
 };
 </script>
