+++ FrontEnd/src/views/Ngo/DonorsList.vue
@@ -3,12 +3,12 @@
     <div id="x-contest" class="container-fluid p-3">
       <div class="row">
         <div class="col-12 p-2">
-          <h3 class="float-start ps-2">NGOs</h3>
-          <router-link to="/addngo" custom v-slot="{ navigate }">
+          <h3 class="float-start ps-2">Donors</h3>
+          <router-link to="/adddonor" custom v-slot="{ navigate }">
             <button
               type="button"
               class="btn btn-primary float-end mx-2"
-              data-nav="admin.ngo.new"
+              data-nav="admin.donor.new"
               @click="navigate"
               role="link"
             >
@@ -20,32 +20,30 @@
         <table class="col-12 table table-striped" id="tbl_ch">
           <thead class="table-dark">
             <tr>
-              <th scope="col">NGO Name</th>
-              <th scope="col">Contact Person</th>
-              <th scope="col">Contact EMail</th>
-              <th scope="col">Contact Password</th>
-              <th scope="col">Contact Phone</th>
+              <th scope="col">Name</th>
+              <th scope="col">Contact</th>
+              <th scope="col">Email</th>
+              <th scope="col">Phone No</th>
               <th scope="col" class="text-center">Actions</th>
             </tr>
           </thead>
           <tbody>
-            <div v-show="!isNgoAvailable">
+            <div v-show="!isDonorsAvailable">
               <tr>
                 <td>No records available</td>
               </tr>
             </div>
-            <tr v-for="item of ngos" :key="item.id">
-              <td>{{ item.ngoName }}</td>
-              <td>{{ item.contactPersonName }}</td>
-              <td>{{ item.contactPersonEmail }}</td>
-              <td>{{ item.contactPersonPassword }}</td>
-              <td>{{ item.contactPersonPhone }}</td>
+            <tr v-for="item of donors" :key="item.id">
+              <td>{{ item.name }}</td>
+              <td>{{ item.contactperson }}</td>
+              <td>{{ item.email }}</td>
+              <td>{{ item.phone }}</td>
               <td class="col-2">
                 <button
                   type="button"
                   class="btn btn-primary float-end mx-2"
                   data-nav="admin.exercise.del"
-                  @click="deleteNgo(item.id)"
+                  @click="deleteDonor(item.id)"
                 >
                   <i class="bi bi-trash"></i>
                 </button>
@@ -53,7 +51,7 @@
                   type="button"
                   class="btn btn-primary float-end mx-2"
                   data-nav="admin.exercise.edit"
-                  @click="editNgo(item.id)"
+                  @click="editDonor(item.id)"
                 >
                   <i class="bi bi-pencil-square"></i>
                 </button>
@@ -86,40 +84,40 @@
 </style>
 <script setup>
 import { onMounted, computed } from "vue";
-import { useNgoStore } from "../stores/ngo";
+import { useDonorsStore } from "../../stores/donors";
 import swal from "sweetalert";
-import router from "../router";
+import router from "../../router";
 
-const store = useNgoStore();
+const store = useDonorsStore();
 
-const ngos = computed(() => {
-  return store.ngos.data;
+const donors = computed(() => {
+  return store.donors.data;
 });
-const isNgoAvailable = computed(() => {
-  return store.ngos?.data?.length > 0;
+const isDonorsAvailable = computed(() => {
+  return store.donors?.data?.length > 0;
 });
-const deleteNgo = (id) => {
+const deleteDonor = (id) => {
   swal({
     title: "Are you sure?",
-    text: "Once deleted, you will not be able to recover this Ngo!",
+    text: "Once deleted, you will not be able to recover this Donor!",
     icon: "warning",
     buttons: true,
     dangerMode: true,
   }).then((willDelete) => {
     if (willDelete) {
-      store.deleteNgo(id);
-      swal("Ngo has been deleted!", {
+      store.deleteDonor(id);
+      swal("Donor has been deleted!", {
         icon: "success",
       });
-      ngos.value = store.fetchNgos();
+      donors.value = store.fetchDonors();
     }
   });
 };
-const editNgo = (id) => {
-  console.log("edit clicked" + id);
-  router.push("editngo/" + id);
+const editDonor = (id) => {
+  console.log("DonorsList :: edit clicked " + id);
+  router.push("editdonor/" + id);
 };
 onMounted(() => {
-  store.fetchNgos();
+  store.fetchDonors();
 });
 </script>
