+++ FrontEnd/src/views/Ngo/EditProgram.vue
@@ -1,19 +1,11 @@
 <script setup>
-import { ref, onMounted, computed, reactive, watch } from "vue";
-import router from "../router";
-import { useProgramStore } from "../stores/program.js";
+import { watch, onMounted, computed, reactive } from "vue";
+import { useProgramStore } from "../../stores/program";
+import { useRoute } from "vue-router";
+import router from "../../router";
 
 const store = useProgramStore();
-
-let donor = reactive({
-  name: "",
-  _id: null,
-});
-
-let supplement = reactive({
-  name: "",
-  _id: null,
-});
+const route = useRoute();
 
 let supplements = computed(() => {
   return store.supplement;
@@ -23,57 +15,40 @@ let donors = computed(() => {
   return store.donor.data;
 });
 
-let selectedDonorObject = computed(() => {
-  return store.donor.data?.find((item) => item.name === donor.name);
-});
-
-watch(selectedDonorObject, (value) => {
-  if (value) {
-    donor._id = value.id;
-  } else {
-    donor._id = null;
-  }
-});
-
-let selectedSupplementObj = computed(() => {
-  return store.supplement.data?.find((item) => item.name === supplement.name);
-});
-
-watch(selectedSupplementObj, (value) => {
-  if (value) {
-    supplement._id = value.id;
-  } else {
-    supplement._id = null;
-  }
-});
-
-let createProgram = {
+let currentProgram = reactive({
   title: "",
   invite_code: "",
-  donor,
-  donor_name: "",
   supplement_name: "",
-  supplement,
+  donor_name: "",
+  notes: "",
   from_date: "",
   to_date: "",
-  notes: "",
-};
+});
+
+currentProgram = computed(() => {
+  if (store.currentprogram) {
+    return store.currentprogram;
+  } else {
+    return {
+      title: "",
+      invite_code: "",
+      supplement_name: "",
+      donor_name: "",
+      notes: "",
+      from_date: "",
+      to_date: "",
+    };
+  }
+});
 
 onMounted(async () => {
   await store.fetchSupplement();
   await store.fecthDonor();
+  await store.fetchProgramById(route.params.id);
 });
 
-const addProgram = () => {
-  let idObject = {
-    suppId: createProgram.supplement._id,
-    donorId: createProgram.donor._id,
-  };
-  delete createProgram.donor;
-  delete createProgram.supplement;
-  createProgram["supplements_details_id"] = idObject.suppId;
-  createProgram["donor_id"] = idObject.donorId;
-  store.createProgram(createProgram);
+const updatingProgram = async () => {
+  await store.updatingProgram(currentProgram);
   return router.push("/programs");
 };
 </script>
@@ -81,10 +56,10 @@ const addProgram = () => {
 <template>
   <div class="full-div container" style="width: 1280px">
     <div id="x-contest" class="container-fluid p-3">
-      <form @submit.prevent="addProgram">
+      <form @submit.prevent="updatingProgram">
         <div class="row">
           <div class="col-12 p-2">
-            <h3 class="float-start">Add Program</h3>
+            <h3 class="float-start">Updating Program</h3>
             <router-link to="/programs" custom v-slot="{ navigate }">
               <button
                 type="button"
@@ -114,7 +89,7 @@ const addProgram = () => {
                 class="form-control"
                 id="exampleFormControlInput1"
                 placeholder="Title"
-                v-model="createProgram.title"
+                v-model="currentProgram.title"
               />
             </div>
 
@@ -125,7 +100,7 @@ const addProgram = () => {
                 class="form-control"
                 id="exampleFormControlInput1"
                 placeholder="From Date"
-                v-model="createProgram.from_date"
+                v-model="currentProgram.from_date"
               />
             </div>
             <div class="col-3">
@@ -135,16 +110,16 @@ const addProgram = () => {
                 class="form-control"
                 id="exampleFormControlInput1"
                 placeholder="To Date"
-                v-model="createProgram.to_date"
+                v-model="currentProgram.to_date"
               />
             </div>
           </div>
           <div class="row mt-4">
             <div class="col-4">
               <label for="exampleFormControlInput1">Donor</label>
-              <select id="level" class="form-select" v-model="donor.name">
+              <select id="level" class="form-select" v-model="currentProgram.donor_name">
                 <option value="" selected>Donor</option>
-                <option v-for="item in donors" v-bind:value="item.name">
+                <option v-for="item in donors" :value="item.name">
                   {{ item.name }}
                 </option>
               </select>
@@ -152,7 +127,11 @@ const addProgram = () => {
 
             <div class="col-4">
               <label for="exampleFormControlInput1">Supplement</label>
-              <select id="level" class="form-select" v-model="supplement.name">
+              <select
+                id="level"
+                class="form-select"
+                v-model="currentProgram.supplement_name"
+              >
                 <option value="" selected>Supplement</option>
                 <option v-for="item in supplements.data" v-bind:value="item.name">
                   {{ item.name }}
@@ -167,14 +146,14 @@ const addProgram = () => {
                 class="form-control"
                 id="exampleFormControlInput1"
                 placeholder="Invite Code"
-                v-model="createProgram.invite_code"
+                v-model="currentProgram.invite_code"
               />
             </div>
           </div>
           <div class="row mt-4">
             <div class="col-12">
               <label for="exampleFormControlInput1">Notes</label>
-              <textarea class="form-control" v-model="createProgram.notes"></textarea>
+              <textarea class="form-control" v-model="currentProgram.notes"></textarea>
             </div>
           </div>
         </div>
@@ -182,4 +161,3 @@ const addProgram = () => {
     </div>
   </div>
 </template>
-<style></style>
