+++ FrontEnd/src/views/Ngo/SupplementList.vue
@@ -3,12 +3,12 @@
     <div id="x-contest" class="container-fluid p-3">
       <div class="row">
         <div class="col-12 p-2">
-          <h3 class="float-start ps-2">All Aanganwadies</h3>
-          <router-link to="/addaanganwadi" custom v-slot="{ navigate }">
+          <h3 class="float-start ps-2">Supplements</h3>
+          <router-link to="/addsupplement" custom v-slot="{ navigate }">
             <button
               type="button"
               class="btn btn-primary float-end mx-2"
-              data-nav="admin.aanganwadi.new"
+              data-nav="admin.supplements.new"
               @click="navigate"
               role="link"
             >
@@ -20,38 +20,34 @@
         <table class="col-12 table table-striped" id="tbl_ch">
           <thead class="table-dark">
             <tr>
-              <th scope="col">Contact Person</th>
-              <th scope="col">Contact Phone</th>
-              <th scope="col">Location</th>
-              <th scope="col">GeoCoordinates</th>
+              <th scope="col">Name</th>
+              <th scope="col">Description</th>
               <th scope="col" class="text-center">Actions</th>
             </tr>
           </thead>
           <tbody>
-            <div v-show="!isAanganwadiAvailable">
+            <div v-show="!isSupplementsAvailable">
               <tr>
                 <td>No records available</td>
               </tr>
             </div>
-            <tr v-for="item of aanganwadies" :key="item.id">
-              <td>{{ item.contactPersonName }}</td>
-              <td>{{ item.contactPersonPhone }}</td>
-              <td>{{ item.location }}</td>
-              <td>{{ item.location_coordinates }}</td>
+            <tr v-for="item of supplements" :key="item.id">
+              <td>{{ item.name }}</td>
+              <td>{{ item.description }}</td>
               <td class="col-2">
                 <button
                   type="button"
                   class="btn btn-primary float-end mx-2"
                   data-nav="admin.exercise.del"
-                  @click="deleteAanganwadi(item.id)"
+                  @click="deleteSupplement(item.id)"
                 >
                   <i class="bi bi-trash"></i>
                 </button>
                 <button
                   type="button"
                   class="btn btn-primary float-end mx-2"
-                  data-nav='admin.exercise.edit?memid="{{item.id}}"'
-                  @click="editAanganwadiPage(item.id)"
+                  data-nav="admin.exercise.edit"
+                  @click="editSupplementPage(item.id)"
                 >
                   <i class="bi bi-pencil-square"></i>
                 </button>
@@ -84,50 +80,40 @@
 </style>
 <script setup>
 import { onMounted, computed } from "vue";
-import { useAanganwadiStore } from "../stores/aanganwadi";
+import { useSupplementStore } from "../../stores/supplement";
 import swal from "sweetalert";
-import router from "../router";
+import router from "../../router";
 
-const store = useAanganwadiStore();
+const store = useSupplementStore();
 
-const aanganwadies = computed(() => {
-  return store.aanganwadies.data;
+const supplements = computed(() => {
+  return store.supplements.data;
 });
-const isAanganwadiAvailable = computed(() => {
-  console.log("aanganwadies -> " + JSON.stringify(store.aanganwadies));
-  return store.aanganwadies != null;
+const isSupplementsAvailable = computed(() => {
+  return store.supplements?.data?.length > 0;
 });
-const deleteAanganwadi = (id) => {
+const deleteSupplement = (id) => {
   swal({
     title: "Are you sure?",
-    text: "Once deleted, you will not be able to recover this Aanganwadi!",
+    text: "Once deleted, you will not be able to recover this Supplement!",
     icon: "warning",
     buttons: true,
     dangerMode: true,
   }).then((willDelete) => {
     if (willDelete) {
-      store.deleteAanganwadi(id);
-      swal("Aanganwadi has been deleted!", {
+      store.deleteSupplement(id);
+      swal("Supplement has been deleted!", {
         icon: "success",
       });
-      store.fetchAanganwadies();
+      supplements.value = store.fetchSupplements();
     }
   });
 };
-
-const editAanganwadiPage = (id) => {
-  router.replace({ path: "/editaanganwadipage/" + id });
+const editSupplementPage = (id) => {
+  console.log("edit clicked" + id);
+  router.push("editsupplement/" + id);
 };
-
-const editAanganwadi = (id) => {
-  store.editAanganwadi(id);
-  swal("Aanganwadi has been updated!", {
-    icon: "success",
-  });
-  store.fetchAanganwadies();
-};
-
 onMounted(() => {
-  store.fetchAanganwadies();
+  store.fetchSupplements();
 });
 </script>
